src/requirements.txt
autopep8==1.5.6
click==7.1.2
jsonlines==2.0.0
neo4j==4.2.1
prettytable==2.1.0
pycodestyle==2.7.0
pymgclient==0.1.0
pytz==2021.1
redis==3.5.3
redisgraph==2.3.0
redisgraph-bulk-loader==0.9.6
toml==0.10.2
wcwidth==0.2.5